pip_requirements.txt
requests==2.22.0
psycopg2==2.8.2
pycodestyle==2.5.0
Django==2.2.1
djangorestframework==3.9.4
gunicorn==19.9.0
whitenoise==4.1.2
dj-database-url==0.5.0
django-heroku==0.3.1
google-auth==1.6.3
